Move relayer specific module events code into relayer crate
