+++ modules/src/core/ics03_connection/events.rs
@@ -4,72 +4,17 @@ use serde_derive::{Deserialize, Serialize};
 use tendermint::abci::tag::Tag;
 use tendermint::abci::Event as AbciEvent;
 
-use crate::core::ics02_client::error::Error as Ics02Error;
 use crate::core::ics02_client::height::Height;
-use crate::core::ics03_connection::error::Error;
 use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
 use crate::events::{IbcEvent, IbcEventType};
 use crate::prelude::*;
 
 /// The content of the `key` field for the attribute containing the connection identifier.
-const HEIGHT_ATTRIBUTE_KEY: &str = "height";
-const CONN_ID_ATTRIBUTE_KEY: &str = "connection_id";
-const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
-const COUNTERPARTY_CONN_ID_ATTRIBUTE_KEY: &str = "counterparty_connection_id";
-const COUNTERPARTY_CLIENT_ID_ATTRIBUTE_KEY: &str = "counterparty_client_id";
-
-pub fn try_from_tx(event: &tendermint::abci::Event) -> Option<IbcEvent> {
-    match event.type_str.parse() {
-        Ok(IbcEventType::OpenInitConnection) => extract_attributes_from_tx(event)
-            .map(OpenInit::from)
-            .map(IbcEvent::OpenInitConnection)
-            .ok(),
-        Ok(IbcEventType::OpenTryConnection) => extract_attributes_from_tx(event)
-            .map(OpenTry::from)
-            .map(IbcEvent::OpenTryConnection)
-            .ok(),
-        Ok(IbcEventType::OpenAckConnection) => extract_attributes_from_tx(event)
-            .map(OpenAck::from)
-            .map(IbcEvent::OpenAckConnection)
-            .ok(),
-        Ok(IbcEventType::OpenConfirmConnection) => extract_attributes_from_tx(event)
-            .map(OpenConfirm::from)
-            .map(IbcEvent::OpenConfirmConnection)
-            .ok(),
-        _ => None,
-    }
-}
-
-fn extract_attributes_from_tx(event: &tendermint::abci::Event) -> Result<Attributes, Error> {
-    let mut attr = Attributes::default();
-
-    for tag in &event.attributes {
-        let key = tag.key.as_ref();
-        let value = tag.value.as_ref();
-        match key {
-            HEIGHT_ATTRIBUTE_KEY => {
-                attr.height = value.parse().map_err(|e| {
-                    Error::ics02_client(Ics02Error::invalid_string_as_height(value.to_string(), e))
-                })?;
-            }
-            CONN_ID_ATTRIBUTE_KEY => {
-                attr.connection_id = value.parse().ok();
-            }
-            CLIENT_ID_ATTRIBUTE_KEY => {
-                attr.client_id = value.parse().map_err(Error::invalid_identifier)?;
-            }
-            COUNTERPARTY_CONN_ID_ATTRIBUTE_KEY => {
-                attr.counterparty_connection_id = value.parse().ok();
-            }
-            COUNTERPARTY_CLIENT_ID_ATTRIBUTE_KEY => {
-                attr.counterparty_client_id = value.parse().map_err(Error::invalid_identifier)?;
-            }
-            _ => {}
-        }
-    }
-
-    Ok(attr)
-}
+pub const HEIGHT_ATTRIBUTE_KEY: &str = "height";
+pub const CONN_ID_ATTRIBUTE_KEY: &str = "connection_id";
+pub const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
+pub const COUNTERPARTY_CONN_ID_ATTRIBUTE_KEY: &str = "counterparty_connection_id";
+pub const COUNTERPARTY_CLIENT_ID_ATTRIBUTE_KEY: &str = "counterparty_client_id";
 
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Attributes {
@@ -292,45 +237,6 @@ impl From<OpenConfirm> for AbciEvent {
         AbciEvent {
             type_str: IbcEventType::OpenConfirmConnection.as_str().to_string(),
             attributes,
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn connection_event_to_abci_event() {
-        let height = Height::new(1, 1).unwrap();
-        let attributes = Attributes {
-            height,
-            connection_id: Some("test_connection".parse().unwrap()),
-            client_id: "test_client".parse().unwrap(),
-            counterparty_connection_id: Some("counterparty_test_conn".parse().unwrap()),
-            counterparty_client_id: "counterparty_test_client".parse().unwrap(),
-        };
-        let mut abci_events = vec![];
-        let open_init = OpenInit::from(attributes.clone());
-        abci_events.push(AbciEvent::from(open_init.clone()));
-        let open_try = OpenTry::from(attributes.clone());
-        abci_events.push(AbciEvent::from(open_try.clone()));
-        let open_ack = OpenAck::from(attributes.clone());
-        abci_events.push(AbciEvent::from(open_ack.clone()));
-        let open_confirm = OpenConfirm::from(attributes);
-        abci_events.push(AbciEvent::from(open_confirm.clone()));
-
-        for event in abci_events {
-            match try_from_tx(&event) {
-                Some(e) => match e {
-                    IbcEvent::OpenInitConnection(e) => assert_eq!(e.0, open_init.0),
-                    IbcEvent::OpenTryConnection(e) => assert_eq!(e.0, open_try.0),
-                    IbcEvent::OpenAckConnection(e) => assert_eq!(e.0, open_ack.0),
-                    IbcEvent::OpenConfirmConnection(e) => assert_eq!(e.0, open_confirm.0),
-                    _ => panic!("unexpected event type"),
-                },
-                None => panic!("converted event was wrong"),
-            }
         }
     }
 }
