+++ relayer/src/chain/cosmos/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod account;
 pub mod config;
+pub mod events;
 pub mod gas;
 pub mod tx;
